+++ src/core/analysis/mcmc/output/TreeTrace.h
@@ -29,7 +29,7 @@ namespace RevBayesCore {
         treeType&                   objectAt(size_t index)                         { return values.at(index); }
         void                        removeLastObject();
         void                        removeObjectAtIndex(int index);
-        int                         size() { return (int)values.size(); }
+        size_t                      size() { return values.size(); }
     
     
         // getters and setters
