+++ src/core/math/distributions/DistributionMultinomial.h
@@ -34,7 +34,7 @@ namespace RevBayesCore {
             double                      lnPdf(const std::vector<double>& p, const std::vector<int>& x);         /*!< Multinomial(p[]) log_e probability */
             double                      lnPdf(const std::vector<double>& p, const std::vector<double>& x);      /*!< Multinomial(p[]) log_e probability */
             std::vector<int>            rv(const std::vector<double>& p, RandomNumberGenerator& rng);           /*!< Multinomial(p[]) random variable */
-            std::vector<int>            rv(const std::vector<double>& p, int n, RandomNumberGenerator& rng);    /*!< Multinomial(p[]) random variable */
+            std::vector<int>            rv(const std::vector<double>& p, size_t n, RandomNumberGenerator& rng);    /*!< Multinomial(p[]) random variable */
         }
     }
 }
